+++ spec/www/spec/db-tx-string-test.js
@@ -349,8 +349,8 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + "INLINE CR-LF String test", function(done) {
-          var db = openDatabase("Inline-CR-LF-String-test.db", "1.0", "Demo", DEFAULT_SIZE);
+        it(suiteName + 'INLINE CR-LF String test', function(done) {
+          var db = openDatabase('INLINE-CR-LF-String-test.db');
           expect(db).toBeDefined();
 
           db.transaction(function(tx) {
@@ -403,8 +403,8 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + "INLINE string tab test", function(done) {
-          var db = openDatabase("Inline-string-tab-test.db", "1.0", "Demo", DEFAULT_SIZE);
+        it(suiteName + 'string tab test (inline vs argument parameter value)', function(done) {
+          var db = openDatabase('string-tab-test.db');
           expect(db).toBeDefined();
 
           db.transaction(function(tx) {
@@ -441,11 +441,11 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + "INLINE string vertical tab test", function(done) {
+        it(suiteName + 'string vertical tab test (inline vs argument parameter value)', function(done) {
           // if (isWP8) pending('...'); // XXX GONE
           if (!isWebSql && !isWindows && isAndroid && !isImpl2) pending('XXX SKIP: BUG on Android (default Android-sqlite-evcore-native-driver access implementation)'); // XXX ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#28
 
-          var db = openDatabase("String-vertical-tab-test.db", "1.0", "Demo", DEFAULT_SIZE);
+          var db = openDatabase('String-vertical-tab-test.db');
           expect(db).toBeDefined();
 
           db.transaction(function(tx) {
@@ -482,11 +482,11 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + "INLINE string form feed test", function(done) {
+        it(suiteName + 'string form feed test (inline vs argument parameter value)', function(done) {
           // if (isWP8) pending('...'); // XXX GONE
           if (!isWebSql && !isWindows && isAndroid && !isImpl2) pending('XXX SKIP: BUG on Android (default Android-sqlite-evcore-native-driver access implementation)'); // XXX ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#28
 
-          var db = openDatabase("String-form-feed-test.db", "1.0", "Demo", DEFAULT_SIZE);
+          var db = openDatabase('String-form-feed-test.db');
           expect(db).toBeDefined();
 
           db.transaction(function(tx) {
@@ -523,11 +523,11 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + "INLINE string backspace test", function(done) {
+        it(suiteName + 'string backspace test (inline vs argument parameter value)', function(done) {
           // if (isWP8) pending('...'); // XXX GONE
           if (!isWebSql && !isWindows && isAndroid && !isImpl2) pending('XXX SKIP: BUG on Android (default Android-sqlite-evcore-native-driver access implementation)'); // XXX ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#28
 
-          var db = openDatabase("String-backspace-test.db", "1.0", "Demo", DEFAULT_SIZE);
+          var db = openDatabase('String-backspace-test.db');
           expect(db).toBeDefined();
 
           db.transaction(function(tx) {
@@ -594,8 +594,7 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'string parameter value manipulation test with UTF-8 2-octet character é', function(done) {
-          if (isAndroid && !isWebSql && !isImpl2) pending('XXX SKIP: BUG on Android (default Android-sqlite-evcore-native-driver access implementation)'); // XXX ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
-
+          // ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
           var db = openDatabase("UTF8-2-octet-upper-value-string-test.db", "1.0", "Demo", DEFAULT_SIZE);
 
           db.transaction(function(tx) {
@@ -650,8 +649,8 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'string parameter value manipulation test with UTF-8 3-octet character €', function(done) {
+          // ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
           if (isWP8) pending('SKIP for WP(8)');
-          if (isAndroid && !isWebSql && !isImpl2) pending('XXX SKIP: BUG on Android (default Android-sqlite-evcore-native-driver access implementation)'); // XXX ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
 
           var db = openDatabase("UTF8-3-octet-string-upper-value-test.db", "1.0", "Demo", DEFAULT_SIZE);
 
@@ -1004,8 +1003,8 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'Inline string manipulation test with UTF-8 2/3 octet characters', function(done) {
+          // ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
           if (isWP8) pending('SKIP for WP(8)');
-          if (isAndroid && !isWebSql && !isImpl2) pending('XXX SKIP: BUG on Android (default Android-sqlite-evcore-native-driver access implementation)'); // XXX ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
 
           var db = openDatabase("Inline-UTF8-string-manipulation-test.db", "1.0", "Demo", DEFAULT_SIZE);
 
@@ -1033,8 +1032,8 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'string parameter manipulation test with UTF-8 2/3 octet characters', function(done) {
+          // ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
           if (isWP8) pending('SKIP for WP(8)');
-          if (isAndroid && !isWebSql && !isImpl2) pending('XXX SKIP: BUG on Android (default Android-sqlite-evcore-native-driver access implementation)'); // XXX ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
 
           var db = openDatabase("UTF8-string-parameter-manipulation-test.db", "1.0", "Demo", DEFAULT_SIZE);