+++ CHANGES.md
@@ -1,5 +1,22 @@
 # Changes
 
+# cordova-sqlite-evplus-ext-free 0.3.1
+
+## cordova-sqlite-evcore-extbuild-free 0.9.7
+
+- fix for multi-byte UTF-8 characters on Android ref:
+  - litehelpers/Cordova-sqlite-evcore-extbuild-free#19 (<https://github.com/litehelpers/Cordova-sqlite-evcore-extbuild-free/issues/19>)
+  - litehelpers/Android-sqlite-evcore-native-driver-free#1 (<https://github.com/litehelpers/Android-sqlite-evcore-native-driver-free/pull/1>)
+  - litehelpers/Android-sqlite-evcore-native-driver-free#2 (<https://github.com/litehelpers/Android-sqlite-evcore-native-driver-free/pull/2>)
+
+### cordova-sqlite-storage 2.2.1
+
+- Fix Android/iOS src copyright, perpetually
+
+## cordova-sqlite-evcore-extbuild-free 0.9.6
+
+- SQLite 3.22.0, with SQLITE_DEFAULT_SYNCHRONOUS=3 (EXTRA DURABLE ref: litehelpers/Cordova-sqlite-storage#736) & other build fixes
+
 # cordova-sqlite-evplus-ext-free 0.3.0
 
 ### cordova-sqlite-evcore-legacy-ext-common-free 0.0.4